# ideas and structures/openapi.frontend.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DAMN 前端所需的后端 API",
    "version": "0.2.0",
    "description": "自适应任务规划器前端（frontend/）当前依赖与将要依赖的全部 HTTP 接口。\n\n约定：\n- API 前缀 `/api/v1`，健康检查在根路径 `/health`，**不在** `/api/v1` 下。\n- 除 `/health`、`/auth/*` 外都需要 `Authorization: Bearer <JWT>`。\n- 所有错误统一返回 `{ code, message, detail }`。\n- 每个操作带 `x-implementation-status`：`existing` = 后端已有；`to-add` = 需要后端新增。\n- 日期用 `YYYY-MM-DD`；时间用 `HH:MM:SS`；时间戳用 ISO 8601。",
    "contact": { "name": "前端 | docs/design/backend-api-gaps.md" }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境（局域网）"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "读取当前用户",
        "description": "头像/昵称/执行权重/自由 profile。",
        "x-implementation-status": "existing",
        "responses": {
          "200": {
            "description": "当前用户",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "更新当前用户",
        "description": "设置页改执行权重、昵称、profile。注意 profile 是全量覆盖语义，前端会先合并整份再提交。",
        "x-implementation-status": "existing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的用户",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/users/me/preferences": {
      "get": {
        "tags": ["users"],
        "summary": "读取排程偏好",
        "description": "设置页「排程偏好」+ 画像二级页「作息」。让「设一次就一直生效」跨设备成立。",
        "x-implementation-status": "to-add",
        "responses": {
          "200": {
            "description": "排程偏好",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SchedulingPreferences" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "保存排程偏好",
        "description": "整体覆盖保存。保存后 `POST /plans/generate` 在请求未显式传这些字段时应回退读取本偏好。",
        "x-implementation-status": "to-add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SchedulingPreferences" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存后的偏好",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SchedulingPreferences" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/v1/users/me/situation/trends": {
      "get": {
        "tags": ["users"],
        "summary": "用户状态趋势（画像）",
        "description": "画像页一级页的曲线：近 N 天的精力/压力/效能，以及数据是否足够、系统判断的理由。",
        "x-implementation-status": "to-add",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "description": "取最近多少天，默认 14",
            "schema": { "type": "integer", "minimum": 2, "maximum": 90, "default": 14 }
          }
        ],
        "responses": {
          "200": {
            "description": "趋势",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SituationTrendRead" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/api/v1/plans/generate": {
      "post": {
        "tags": ["plans"],
        "summary": "生成计划（同步）",
        "description": "由目标直接生成并落库。前端的「拆解确认」在后端 draft 接口就绪前会回退到这里。同步返回完整 plan，不要做 SSE 假进度。",
        "x-implementation-status": "existing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/PlanGenerateRequest" } }
          }
        },
        "responses": {
          "202": {
            "description": "生成完成（含 plan）",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PlanGenerateResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "未认证（401）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
...
      "Validation": {
        "description": "参数校验失败（422）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "description": "统一错误体",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码，如 `not_found` / `conflict` / `replan_not_eligible` / `validation_error`"
          },
          "message": { "type": "string", "description": "给人看的说明" },
          "detail": { "nullable": true, "description": "可选的补充信息（对象或数组）" }
        }
      },
...
      "UserRead": {
        "type": "object",
        "required": ["id", "email", "execution_weight", "profile", "created_at"],
        "properties": {
          "id": { "type": "integer" },
          "email": { "type": "string", "format": "email" },
          "display_name": { "type": "string", "nullable": true },
          "execution_weight": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "执行权重，前端设置页可调；用于预估与画像「效能」"
          },
          "profile": {
            "type": "object",
            "additionalProperties": true,
            "description": "自由 JSON；建议存放 scheduling_preferences（见下）"
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "UserUpdate": {
        "type": "object",
        "properties": {
          "display_name": { "type": "string", "nullable": true },
          "execution_weight": { "type": "number", "minimum": 0, "maximum": 1 },
          "profile": {
            "type": "object",
            "additionalProperties": true,
            "description": "全量覆盖语义：请提交完整 profile"
          }
        }
      },
      "SchedulingPreferences": {
        "type": "object",
        "description": "设置页「排程偏好」与画像二级页「基础档案」共用的数据",
        "required": [
          "available_minutes_per_day",
          "daily_limit_minutes",
          "buffer_minutes",
          "high_cognitive_max_per_day"
        ],
        "properties": {
          "available_minutes_per_day": { "type": "integer", "minimum": 30, "maximum": 1440, "default": 480 },
          "daily_limit_minutes": { "type": "integer", "minimum": 30, "maximum": 1440, "default": 300 },
          "buffer_minutes": { "type": "integer", "minimum": 0, "maximum": 120, "default": 15 },
          "high_cognitive_max_per_day": { "type": "integer", "minimum": 1, "maximum": 10, "default": 2 },
          "sleep_start": { "type": "string", "example": "23:30", "description": "本地时间 HH:MM" },
          "sleep_end": { "type": "string", "example": "07:30", "description": "本地时间 HH:MM" }
        }
      },
      "TaskStatus": {
        "type": "string",
        "enum": ["pending", "scheduled", "in_progress", "completed", "skipped", "failed"]
      },
      "CognitiveLoad": {
        "type": "string",
        "enum": ["low", "medium", "high", "restorative"]
      },
      "StandardRead": {
        "type": "object",
        "required": ["id", "description", "completed", "order_index"],
        "properties": {
          "id": { "type": "integer" },
          "description": { "type": "string" },
          "estimated_duration": { "type": "integer", "nullable": true },
          "completed": { "type": "boolean" },
          "order_index": { "type": "integer" }
        }
      },
      "TaskRead": {
        "type": "object",
        "description": "前端时间轴依赖 scheduled_date / start_time / end_time / priority",
        "required": ["id", "plan_id", "title", "estimated_duration", "priority", "status", "order_index"],
        "properties": {
          "id": { "type": "integer" },
          "plan_id": { "type": "integer" },
          "goal_id": { "type": "integer", "nullable": true, "description": "可信；用于待办按目标分组" },
          "parent_task_id": { "type": "integer", "nullable": true },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "estimated_duration": { "type": "integer", "description": "预计分钟" },
          "predicted_duration": { "type": "integer", "nullable": true },
          "cognitive_load": { "$ref": "#/components/schemas/CognitiveLoad" },
          "priority": {
            "type": "integer",
            "description": "1=低 2=中 3=高（前端旗标按 >=3 高 / 2 中 / 其它 低）。请给出真实档位差异。"
          },
          "scheduled_date": { "type": "string", "format": "date", "nullable": true },
          "start_time": { "type": "string", "nullable": true, "example": "09:00:00" },
          "end_time": { "type": "string", "nullable": true, "example": "10:00:00" },
          "status": { "$ref": "#/components/schemas/TaskStatus" },
          "completion_probability": { "type": "number", "nullable": true },
          "is_flexible": { "type": "boolean" },
          "order_index": { "type": "integer" },
          "standards": { "type": "array", "items": { "$ref": "#/components/schemas/StandardRead" } }
        }
      },
      "GoalRead": {
        "type": "object",
        "required": ["id", "title", "goal_type", "priority", "status"],
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string" },
          "goal_type": { "type": "string", "enum": ["long_term", "short_term", "project", "habit", "other"] },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4 },
          "status": { "type": "string" }
        }
      },
...
      "GoalCreate": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "goal_type": { "type": "string", "enum": ["long_term", "short_term", "project", "habit", "other"], "default": "other" },
          "deadline": { "type": "string", "format": "date-time", "nullable": true },
          "priority": { "type": "integer", "minimum": 1, "maximum": 4, "default": 2 },
          "estimated_minutes": { "type": "integer", "nullable": true },
          "subject": { "type": "string", "nullable": true },
          "task_type": { "type": "string", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["active", "draft"],
            "default": "active",
            "description": "建待拆解项时传 draft"
          }
        }
      },
...
      "PlanRead": {
        "type": "object",
        "required": ["id", "user_id", "version", "status", "start_date", "end_date", "created_at", "tasks", "goals"],
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "version": { "type": "integer" },
          "status": { "type": "string", "enum": ["active", "superseded", "draft", "completed", "archived"] },
          "title": { "type": "string", "nullable": true },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "parent_plan_id": { "type": "integer", "nullable": true },
          "confidence": { "type": "number", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "tasks": { "type": "array", "items": { "$ref": "#/components/schemas/TaskRead" } },
          "goals": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GoalRead" },
            "description": "由该计划任务的 goal_id 反查；前端只用它做分组标题"
          }
        }
      },
...
      "PlanGenerateRequest": {
        "type": "object",
        "required": ["goals"],
        "properties": {
          "goals": { "type": "array", "minItems": 1, "items": { "$ref": "#/components/schemas/GoalCreate" } },
          "start_date": { "type": "string", "format": "date", "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true },
          "plan_title": { "type": "string", "nullable": true },
          "available_minutes_per_day": { "type": "integer", "minimum": 1, "maximum": 1440, "default": 480 },
          "daily_limit_minutes": { "type": "integer", "minimum": 1, "maximum": 1440, "default": 300 },
          "buffer_minutes": { "type": "integer", "minimum": 0, "maximum": 120, "default": 15 },
          "high_cognitive_max_per_day": { "type": "integer", "minimum": 1, "maximum": 10, "default": 2 },
          "user_profile": { "type": "object", "additionalProperties": true, "default": {} },
          "execution_weight": { "type": "number", "nullable": true }
        }
      },
      "PlanGenerateResponse": {
        "type": "object",
        "required": ["job_id", "status", "events_url"],
        "properties": {
          "job_id": { "type": "string" },
          "status": { "type": "string", "example": "completed" },
          "plan_id": { "type": "integer", "nullable": true },
          "events_url": { "type": "string" },
          "plan": { "allOf": [{ "$ref": "#/components/schemas/PlanRead" }], "nullable": true }
        }
      },
...
      "SituationTrendPoint": {
        "type": "object",
        "required": ["date"],
        "properties": {
          "date": { "type": "string", "format": "date" },
          "energy": { "type": "number", "nullable": true, "description": "0..10" },
          "stress": { "type": "number", "nullable": true, "description": "0..10" },
          "efficacy": { "type": "number", "nullable": true, "description": "0..1；如无后端定义，前端按 0.6*execution_weight + 0.4*完成率 兜底" }
        }
      },
      "SituationTrendRead": {
        "type": "object",
        "required": ["points", "samples", "min_samples", "sufficient"],
        "properties": {
          "points": { "type": "array", "items": { "$ref": "#/components/schemas/SituationTrendPoint" } },
          "samples": { "type": "integer", "description": "有效反馈天数" },
          "min_samples": { "type": "integer", "description": "判断门槛，建议 5" },
          "sufficient": { "type": "boolean" },
          "drivers": {
            "type": "array",
            "items": { "type": "string" },
            "description": "系统判断理由（人话）"
          }
        }
      }
    }
  }
}